+++ lib/src/chart/line/presenters.dart
@@ -61,46 +61,6 @@ class LineAndHotspotPresenter extends Presenter {
   }
 }
 
-/* todo -2 remove when tested
-  LineAndHotspotPresenter({
-    StackableValuePoint point,
-    StackableValuePoint nextRightColumnValuePoint,
-    int rowIndex,
-    ChartLayouter layouter,
-  })
-      : super(
-    point: point,
-    nextRightColumnValuePoint: nextRightColumnValuePoint,
-    rowIndex: rowIndex,
-    layouter: layouter,
-  ){
-    var options = layouter.options as LineChartOptions;
-
-    rowDataPaint = new ui.Paint();
-    rowDataPaint.color = layouter.data.dataRowsColors[rowIndex % layouter.data.dataRowsColors.length];
-
-    ui.Offset fromPoint = point.scaledTo;
-    ui.Offset toPoint = nextRightColumnValuePoint?.scaledTo;
-    toPoint ??= fromPoint;
-    linePresenter = new line_presenter.LinePresenter(
-      from: fromPoint,
-      to: toPoint,
-      paint: rowDataPaint..strokeWidth = options.lineStrokeWidth,
-    );
-    offsetPoint = fromPoint; // point is the left (from) end of the line
-    //innerPaint = new ui.Paint();
-    //innerPaint.color = material.Colors.yellow;
-    //outerPaint = new ui.Paint();
-    //outerPaint.color = material.Colors.black;
-    innerPaint = options.hotspotInnerPaint;
-    outerPaint = options.hotspotOuterPaint;
-    innerRadius = options.hotspotInnerRadius;
-    outerRadius = options.hotspotOuterRadius;
-  }
-}
-
- */
-
 /// Creator of the [LineAndHotspotPresenter] instances - the leaf visual
 /// elements on the line chart (point and line showing one data value).
 ///